Introduce unified pynx CLI with distributed sub-command modules
Consolidates all CLI entry points under a single `pynx` command group:
- `pynx convert` (data conversion + generate-template, get-readers, reader-info sub-commands)
- `pynx validate` (NeXus file validation)
- `pynx read` (NeXus annotator/debugger)
- `pynx generate-eln` (ELN scaffold generator)
- `pynx inspect-appdef` (application definition inspector)

Each sub-package owns its own `cli.py`; `src/pynxtools/cli.py` assembles them.
Legacy entry points (dataconverter, validate_nexus, read_nexus, generate_eln) remain
as deprecated aliases. Updates all docs and adds CLI smoke tests for pynx read,
inspect-appdef, get-readers, and reader-info.

Fixes mypy errors: correct type: ignore comment codes in nexus.py and
eln_mapper.py, explicit Literal annotation for node_type in validation.py,
and type: ignore[operator] for get_reader()-based instantiation in cli.py.
