+++ .cspell/custom-dictionary.txt
@@ -197,6 +197,7 @@ posint
 posteriori
 printoptions
 punx
+pynx
 pynxtools
 raman
 rdcc
