+++ src/pynxtools/dataconverter/convert.py
@@ -18,23 +18,14 @@
 """This script runs the conversion routine using a selected reader and write out a NeXus file."""
 
 import glob
-import importlib.machinery
 import importlib.util
-import json
 import logging
 import os
-import sys
-from gettext import gettext
-from pathlib import Path
-from typing import Literal
 
-import click
 import lxml.etree as ET
 import yaml
-from click_default_group import DefaultGroup
 
 from pynxtools.dataconverter import helpers
-from pynxtools.dataconverter.nexus_tree import generate_tree_from
 from pynxtools.dataconverter.readers.base.reader import BaseReader
 from pynxtools.dataconverter.template import Template
 from pynxtools.dataconverter.validation import validate_dict_against
@@ -247,256 +238,3 @@ def parse_params_file(params_file):
     for param in list(params.keys()):
         params[param.replace("-", "_")] = params.pop(param)
     return params
-
-
-class CustomClickGroup(DefaultGroup):
-    def format_options(
-        self, ctx: click.Context, formatter: click.HelpFormatter
-    ) -> None:
-        """Writes all the options into the formatter if they exist."""
-        opts = []
-        for param in self.get_params(ctx) + ctx.command.commands["convert"].params:  # type: ignore
-            rv = param.get_help_record(ctx)
-            if rv is not None:
-                opts.append(rv)
-
-        if opts:
-            with formatter.section(gettext("Options")):
-                formatter.write_dl(opts)
-        self.format_commands(ctx, formatter)
-        with formatter.section(gettext("Info")):
-            formatter.write_text(
-                "You can see more options by using --help for specific commands. For example: dataconverter generate-template --help"
-            )
-
-
-@click.group(
-    cls=CustomClickGroup,
-    default="convert",
-    default_if_no_args=True,
-)
-def main_cli():
-    pass
-
-
-@main_cli.command("convert")
-@click.argument("files", nargs=-1, type=click.Path(exists=True))
-@click.option(
-    "--input-file",
-    default=[],
-    multiple=True,
-    help=(
-        "Deprecated: Please use the positional file arguments instead. The path to the "
-        "input data file to read. Repeat for more than one file. default=[] This option "
-        "is required if no '--params-file' is supplied."
-    ),
-)
-@click.option(
-    "--reader",
-    default="json_map",
-    type=click.Choice(get_names_of_all_readers(), case_sensitive=False),
-    help=(
-        "The reader to use. Examples are json_map or readers from a pynxtools plugin. "
-        "default='json_map' This option is required if no '--params-file' is supplied."
-    ),
-)
-@click.option(
-    "--nxdl",
-    default=None,
-    help=(
-        "The name of the NeXus application definition file to use without the extension "
-        "nxdl.xml. This option is required if no '--params-file' is supplied."
-    ),
-)
-@click.option(
-    "--output",
-    default="output.nxs",
-    help="The path to the output NeXus file to be generated. default='output.nxs'",
-)
-@click.option(
-    "--params-file",
-    type=click.File("r"),
-    default=None,
-    help="Allows to pass a .yaml file with all the parameters the converter supports.",
-)
-@click.option(
-    "--append",
-    is_flag=True,
-    default=False,
-    help="Appends instance data to the given <output> file if the file already exists."
-    "Will prevent an overwriting of existent data and adding of new NXentry instances."
-    "If True, will skip the validation whatever irrespective if --skip-verify is passed.",
-)
-@click.option(
-    "--ignore-undocumented",
-    is_flag=True,
-    default=False,
-    help="Ignore all undocumented concepts during validation.",
-)
-@click.option(
-    "--fail",
-    is_flag=True,
-    default=False,
-    help="Fail conversion and don't create an output file if the validation fails.",
-)
-@click.option(
-    "--skip-verify",
-    is_flag=True,
-    default=False,
-    help="Skips the verification routine during conversion."
-    "When --append is used the verification is always skipped"
-    "irrespective if --skip-verify is used or not.",
-)
-@click.option(
-    "--mapping",
-    help=(
-        "Takes a <name>.mapping.json file and converts data from given input files. "
-        "Deprecated. Will be removed in a future release. The --config flag can be used instead."
-    ),
-)
-@click.option(
-    "-c",
-    "--config",
-    "config_file",
-    type=click.Path(exists=True, dir_okay=False, file_okay=True, readable=True),
-    default=None,
-    help="A json config file for the reader",
-)
-# pylint: disable=too-many-arguments
-def convert_cli(
-    files: tuple[str, ...],
-    input_file: tuple[str, ...],
-    reader: str,
-    nxdl: str,
-    output: str,
-    params_file: str,
-    append: bool,
-    ignore_undocumented: bool,
-    skip_verify: bool,
-    config_file: str,
-    fail: bool,
-    mapping: str | None,
-    **kwargs,
-):
-    """This command allows you to use the converter functionality of the dataconverter."""
-    if params_file:
-        try:
-            convert(**parse_params_file(params_file))
-            return
-        except TypeError as exc:
-            sys.tracebacklimit = 0
-            raise click.UsageError(
-                "Please make sure you have the following entries in your "
-                "parameter file:\n\n# NeXusParser Parameter File - v0.0.1"
-                "\n\ndataconverter:\n\treader: value\n\tnxdl: value\n\tin"
-                "put-file: value"
-            ) from exc
-    if nxdl is None:
-        raise click.UsageError("Missing option '--nxdl'")
-    if mapping:
-        logger.warning(
-            "The --mapping option is deprecated. Please use a config file with the -c argument instead."
-        )
-        reader = "json_map"
-        input_file = input_file + tuple([mapping])
-
-    if config_file:
-        kwargs["config_file"] = config_file
-
-    file_list = []
-    for file in files:
-        if os.path.isdir(file):
-            p = Path(file)
-            for f in p.rglob("*"):
-                if f.is_file():
-                    file_list.append(str(f))
-            continue
-        file_list.append(file)
-
-    if input_file:
-        logger.warning(
-            "The --input-file option is deprecated. Please use the positional arguments instead."
-        )
-
-    try:
-        convert(
-            tuple(file_list) + input_file,
-            reader,
-            nxdl,
-            output,
-            skip_verify,
-            append=append,
-            ignore_undocumented=ignore_undocumented,
-            fail=fail,
-            **kwargs,
-        )
-    except FileNotFoundError as exc:
-        raise click.BadParameter(str(exc)) from exc
-    except ValidationFailed as exc:
-        raise click.ClickException(
-            "Validation failed: No file written because '--fail' was requested."
-        ) from exc
-
-
-@main_cli.command()
-@click.option(
-    "--nxdl",
-    default=None,
-    help=(
-        "The name of the NeXus application definition file to use without extension. For example: NXmpes"
-    ),
-    required=True,
-)
-@click.option(
-    "--required",
-    help="Use this flag to only get the required template.",
-    is_flag=True,
-)
-@click.option(
-    "--pythonic",
-    help="Prints a valid Python dictionary instead of JSON",
-    is_flag=True,
-)
-@click.option(
-    "--output",
-    help="Writes the output into the filepath provided.",
-    type=click.Path(),
-)
-def generate_template(nxdl: str, required: bool, pythonic: bool, output: str):
-    "Generates and prints a template to use for your nxdl."
-
-    def write_to_file(text):
-        f = open(output, "w")
-        f.write(text)
-        f.close()
-
-    tree = generate_tree_from(nxdl)
-
-    print_or_write = lambda txt: write_to_file(txt) if output else print(txt)
-
-    level: Literal["required", "recommended", "optional"] = "optional"
-    if required:
-        level = "required"
-    reqs = tree.required_fields_and_attrs_names(level=level)
-    template = {
-        helpers.convert_nxdl_path_dict_to_data_converter_dict(req): None for req in reqs
-    }
-
-    if pythonic:
-        print_or_write(str(template))
-        return
-    print_or_write(
-        json.dumps(
-            template,
-            indent=4,
-            sort_keys=True,
-            ensure_ascii=False,
-        )
-    )
-
-
-@main_cli.command("get-readers")
-def get_reader_cli():
-    """Prints a list of all installed readers."""
-    readers = get_names_of_all_readers()
-    logger.info(f"The following readers are currently installed: {readers}.")
